add sub updaterpmdb
When the packages are extracted into the target, they do not show up
when doing "chroot targetdir rpm -qa". This is because they are only
extracted but not installed. Now call rpm with --justdb which fixes
the rpm db, and all packages are recognized as installed.

Then, much less packages are installed in the postinst part.